backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
psycopg2-binary
python-multipart
insightface
onnxruntime-gpu
minio
pandas
openpyxl
pydantic
pydantic-settings